requirements.txt
fastapi
sqlalchemy
uvicorn
python-dotenv
mysql-connector-python
pymysql